Refactor AddServerScreen to avoid ViewModel in components
