+++ features/auth/ui/src/main/kotlin/com/boswelja/truemanager/auth/ui/addserver/AddServerComponents.kt
@@ -1,24 +1,16 @@
 package com.boswelja.truemanager.auth.ui.addserver
 
-import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
 import androidx.compose.animation.shrinkVertically
-import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.calculateEndPadding
-import androidx.compose.foundation.layout.calculateStartPadding
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
@@ -34,191 +26,16 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.boswelja.truemanager.auth.ui.R
-import kotlinx.coroutines.flow.collectLatest
-import org.koin.androidx.compose.koinViewModel
-
-@Composable
-fun AuthComponents(
-    onLoginSuccess: () -> Unit,
-    modifier: Modifier = Modifier,
-    contentPadding: PaddingValues = PaddingValues(),
-    viewModel: AddServerViewModel = koinViewModel()
-) {
-    val layoutDirection = LocalLayoutDirection.current
-    val isLoading by viewModel.isLoading.collectAsState()
-
-    var serverName by rememberSaveable { mutableStateOf("") }
-    var serverAddress by rememberSaveable { mutableStateOf("") }
-    var selectedAuthType by rememberSaveable(saver = rememberAuthTypeSaver()) {
-        mutableStateOf(AuthTypes.first())
-    }
-
-    var username by rememberSaveable { mutableStateOf("") }
-    var password by rememberSaveable { mutableStateOf("") }
-    var apiKey by rememberSaveable { mutableStateOf("") }
-    var isServerAddressInvalid by rememberSaveable {
-        mutableStateOf(false)
-    }
-    var isUsernameOrPasswordInvalid by rememberSaveable {
-        mutableStateOf(false)
-    }
-    var isApiKeyInvalid by rememberSaveable {
-        mutableStateOf(false)
-    }
-
-    val logIn = {
-        when (selectedAuthType) {
-            AuthType.ApiKeyAuth -> viewModel.tryLogIn(serverName, serverAddress, apiKey)
-            AuthType.BasicAuth -> viewModel.tryLogIn(serverName, serverAddress, username, password)
-        }
-    }
-    val loginEnabled by remember {
-        derivedStateOf {
-            val authValid = when (selectedAuthType) {
-                AuthType.ApiKeyAuth -> apiKey.isNotBlank() && !isApiKeyInvalid
-                AuthType.BasicAuth -> username.isNotBlank() && password.isNotBlank() && !isUsernameOrPasswordInvalid
-            }
-            !isLoading && serverAddress.isNotBlank() && authValid && !isServerAddressInvalid
-        }
-    }
-
-    // TODO handle login properly
-    LaunchedEffect(viewModel) {
-        viewModel.events.collectLatest {
-            when (it) {
-                AddServerViewModel.Event.LoginSuccess -> onLoginSuccess()
-                AddServerViewModel.Event.LoginFailedNotHttps,
-                AddServerViewModel.Event.LoginFailedServerNotFound ->
-                    isServerAddressInvalid = true
-                AddServerViewModel.Event.LoginFailedKeyInvalid ->
-                    isApiKeyInvalid = true
-                AddServerViewModel.Event.LoginFailedUsernameOrPasswordInvalid ->
-                    isUsernameOrPasswordInvalid = true
-                AddServerViewModel.Event.LoginFailedKeyAlreadyExists -> TODO()
-                null -> return@collectLatest
-            }
-            viewModel.clearPendingEvent()
-        }
-    }
-
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        ServerNameField(
-            serverName = serverName,
-            onServerNameChange = { serverName = it },
-            enabled = !isLoading,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    top = contentPadding.calculateTopPadding(),
-                    start = contentPadding.calculateStartPadding(layoutDirection),
-                    end = contentPadding.calculateEndPadding(layoutDirection)
-                )
-                .widthIn(max = 560.dp)
-        )
-        Spacer(Modifier.height(8.dp))
-        ServerAddressField(
-            serverAddress = serverAddress,
-            onServerAddressChange = {
-                serverAddress = it
-                isServerAddressInvalid = false
-            },
-            enabled = !isLoading,
-            error = isServerAddressInvalid,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    start = contentPadding.calculateStartPadding(layoutDirection),
-                    end = contentPadding.calculateEndPadding(layoutDirection)
-                )
-                .widthIn(max = 560.dp)
-        )
-        AuthTypeSelector(
-            currentType = selectedAuthType,
-            onAuthTypeChange = { selectedAuthType = it },
-            modifier = Modifier.padding(vertical = 16.dp),
-            enabled = !isLoading
-        )
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .widthIn(max = 560.dp)
-                .padding(
-                    start = contentPadding.calculateStartPadding(layoutDirection),
-                    end = contentPadding.calculateEndPadding(layoutDirection),
-                    bottom = contentPadding.calculateBottomPadding()
-                )
-        ) {
-            AnimatedContent(
-                targetState = selectedAuthType,
-                label = "Login method content",
-                transitionSpec = { fadeIn() togetherWith fadeOut() },
-            ) { authType ->
-                when (authType) {
-                    AuthType.ApiKeyAuth -> {
-                        ApiKeyFields(
-                            apiKey = apiKey,
-                            onApiKeyChange = {
-                                apiKey = it
-                                isApiKeyInvalid = false
-                            },
-                            onDone = logIn,
-                            enabled = !isLoading,
-                            error = isApiKeyInvalid,
-                            modifier = Modifier.fillMaxWidth()
-                        )
-                    }
-                    AuthType.BasicAuth -> {
-                        BasicAuthFields(
-                            username = username,
-                            onUsernameChange = {
-                                username = it
-                                isUsernameOrPasswordInvalid = false
-                            },
-                            password = password,
-                            onPasswordChange = {
-                                password = it
-                                isUsernameOrPasswordInvalid = false
-                            },
-                            onDone = logIn,
-                            enabled = !isLoading,
-                            error = isUsernameOrPasswordInvalid,
-                            modifier = Modifier.fillMaxWidth()
-                        )
-                    }
-                }
-            }
-            Spacer(modifier = Modifier.height(16.dp))
-            LoginButton(
-                onClick = logIn,
-                enabled = loginEnabled,
-                loading = isLoading,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
 
 @Composable
 fun ServerNameField(
